requirements.txt
argon2-cffi==21.1.0
async-generator @ file:///home/ktietz/src/ci/async_generator_1611927993394/work
attrs==21.2.0
backcall==0.2.0
bleach==4.1.0
certifi==2021.5.30
cffi==1.14.6
cycler==0.10.0
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
entrypoints==0.3
findspark==1.4.2
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1617874469820/work
ipykernel==6.3.1
ipython==7.27.0
ipython-genutils==0.2.0
ipywidgets==7.6.4
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.2
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.1
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
mne==0.23.3
nb-conda-kernels @ file:///tmp/build/80754af9/nb_conda_kernels_1606775941989/work
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.3
numpy==1.21.2
packaging==21.0
pandas==1.3.2
pandocfilters==1.4.3
parso==0.8.2
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare==0.7.5
Pillow==8.3.2
prometheus-client==0.11.0
prompt-toolkit==3.0.20
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-dotenv==0.19.0
pytz==2021.1
pyzmq==22.2.1
qtconsole==5.1.1
QtPy==1.11.0
scipy==1.7.1
seaborn==0.11.2
Send2Trash==1.8.0
six @ file:///tmp/build/80754af9/six_1623709665295/work
terminado==0.12.1
testpath==0.5.0
tornado==6.1
traitlets==5.1.0
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp @ file:///tmp/build/80754af9/zipp_1625570634446/work
